Fix arg type of 'base' to int for strtoll_l and strtoull_l
See bionic 3c5037f1b3b747e79d17a5f717d9f9c365132d33

+++ sources/android/support/src/locale_support.c
@@ -109,11 +109,11 @@ unsigned long strtoul_l(const char *nptr, char **endptr, int base, locale_t loc)
     return strtoul(nptr, endptr, base);
 }
 
-long long strtoll_l(const char *nptr, char **endptr, size_t base, locale_t loc) {
+long long strtoll_l(const char *nptr, char **endptr, int base, locale_t loc) {
     return strtoll(nptr, endptr, base);
 }
 
-unsigned long long strtoull_l(const char *nptr, char **endptr, size_t base, locale_t loc) {
+unsigned long long strtoull_l(const char *nptr, char **endptr, int base, locale_t loc) {
     return strtoull(nptr, endptr, base);
 }
